Support pattern and type synonyms.
It's interesting that the `type` is inferred in `SomeClass(type (&+&))`.
It can't be inferred at top level because there's a risk of clashing
between type constructors and values. I tried putting a value named
`(&+&)` inside `SomeClass`, and then couldn't import either of them
(they apparently weren't members of the class). A.hs still compiled,
though.

Not really sure what's going on there.